requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
distlib==0.3.6
dj-database-url==1.3.0
Django==4.2
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-on-heroku==1.1.2
djangorestframework==3.14.0
filelock==3.11.0
gunicorn==20.1.0
pipenv==2023.3.20
platformdirs==3.2.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
whitenoise==6.4.0
